Improved navigation, added startActivityOnResult. Added several cardviews,  added elevation. Reset database version number. Ran Cleanup. Fixed several bugs. Build APK. Application Finished.

+++ GamesBacklogManager/app/src/main/java/nl/hva/gamesbacklogmanager/model/Game.java
@@ -40,7 +40,7 @@ public class Game implements Serializable {
     }
 
     public long getId() {
-        return id;
+        return (long) id;
     }
 
     public void setId(int id) {
